refactor: profit 계산의 반환 타입을 long 으로 변경해 overflow 를 방지해라
반환타입의 변경
- 만약 우연히 한 고객이 겹치는 중복되는 로또 번호를 배정받고, 1등이 된다면 profit 에 overflow 가 일어나므로 수정한다.

+++ src/main/java/lotto/utils/ProfitCalculator.java
@@ -8,7 +8,7 @@ public class ProfitCalculator {
         return ((double) (calculateTotalProfit(lottoTickets)) / paidAmount) * 100;
     }
 
-    private static int calculateTotalProfit(List<LottoTicket> lottoTickets) {
+    private static long calculateTotalProfit(List<LottoTicket> lottoTickets) {
         int totalProfit = 0;
 
         for (LottoTicket lottoTicket : lottoTickets) {
